# Add `intersect(::AbstractRange, ::AbstractVector)`