Closes #41759