@barucden 2021-08-03
@IanButterworth Cannot refer to `AbstractVector` from `range.jl`. Should I just move the definitions to `array.jl` (originally I thought it is more fitting in `range.jl`)?