@IanButterworth 2021-08-03
Moving to `array.jl` sounds good. Perhaps stylistically you could flip the args so the first arg makes most sense for the new location, but that's very minor
```julia
function intersect(vec::AbstractVector, r::AbstractRange)
    common = Iterators.filter(x -> x ∈ r, vec)
    seen = Set{eltype(vec)}(common)
    return vectorfilter(_shrink_filter!(seen), common)
end
intersect(r::AbstractRange, vec::AbstractVector) = intersect(vec, r)
```