@simeonschaub 2021-08-03
We don't currently define an explicit method for `intersect(::AbstractRange, ::AbstractRange)`, so you will need to define that as well to avoid ambiguities.