@simeonschaub 2021-08-03
> Should it be just a fall-back method with a naive implementation? Like looping through the shorter range and checking if each element is in the longer range?

Yes, seems reasonable. 

> Is there a set implementation which guarantees the insertion order in Julia?

No, not currently. That would be a much bigger project, so I wouldn't bother for now.