@DilumAluthge 2021-08-03
> @DilumAluthge Aren't the (non-freebsd) errors here real ambiguity test failures related to this PR?

Oh you're right. I just assumed they were the Linux `EAI_AGAIN` errors 😂