@barucden 2021-08-04
Rebased and added `intersect(::AbstractRange, ::AbstractRange)`. @simeonschaub Would you check if [this test](https://github.com/JuliaLang/julia/pull/41769/commits/8c2c5237f4fdba3c5269ee725313d0efd2c365c3#diff-1d6e3ecff651f017ed32b54d0919e13362ac278dd18df1cc8b2feda52e630744) is okay, please?