@IanButterworth 2021-08-10
Some return type comparison and benchmarks vs. master
```julia
using BenchmarkTools
for a in [1:4, 1:1.0:4, 10:4:30, collect(1:4)]
    for b in [2:4, 2:1.0:4, 14:4:30, collect(2:6)]
        r = Base.intersect(a, b)
        print(typeof(a), "\t", typeof(b), "\t", typeof(r),"\t", r, "\t")
        @btime Base.intersect($a, $b)
    end
end
```


| a         | b           | Master return   | PR Return  | Master                                  | PR                                     |
|-----------|-------------|-----------------|-----------------|-----------------------------------------|----------------------------------------|
| 1:4       | 2:4         | 2:4             | 2:4             | 0.010 ns (0 allocations: 0 bytes)       | 0.010 ns (0 allocations: 0 bytes)      |
| 1:4       | 2:1.0:4     | [2, 3, 4]       | [2.0, 3.0, 4.0] | 376.512 ns (16 allocations: 1.12 KiB)   | 66.507 ns (2 allocations: 144 bytes)   |
| 1:4       | 14:4:30     | 14:4:13         | 14:4:13         | 0.010 ns (0 allocations: 0 bytes)       | 0.010 ns (0 allocations: 0 bytes)      |
| 1:4       | [2,3,4,5,6] | [2, 3, 4]       | [2, 3, 4]       | 364.560 ns (14 allocations: 1.03 KiB)   | 156.247 ns (7 allocations: 608 bytes)  |
| 1:1.0:4   | 2:4         | [2.0, 3.0, 4.0] | [2.0, 3.0, 4.0] | 422.495 ns (16 allocations: 1.16 KiB)   | 72.938 ns (2 allocations: 144 bytes)   |
| 1:1.0:4   | 2:1.0:4     | [2.0, 3.0, 4.0] | [2.0, 3.0, 4.0] | 416.785 ns (15 allocations: 1.12 KiB)   | 87.419 ns (2 allocations: 144 bytes)   |
| 1:1.0:4   | 14:4:30     | Float64[]       | Float64[]       | 406.565 ns (15 allocations: 1.08 KiB)   | 44.098 ns (1 allocation: 64 bytes)     |
| 1:1.0:4   | [2,3,4,5,6] | [2.0, 3.0, 4.0] | [2, 3, 4]       | 421.245 ns (14 allocations: 1.06 KiB)   | 234.930 ns (7 allocations: 608 bytes)  |
| 10:4:30   | 2:4         | 10:4:9          | 10:4:9          | 0.010 ns (0 allocations: 0 bytes)       | 0.010 ns (0 allocations: 0 bytes)      |
| 10:4:30   | 2:1.0:4     | Int64[]         | Float64[]       | 401.403 ns (16 allocations: 1.12 KiB)   | 41.474 ns (1 allocation: 64 bytes)     |
| 10:4:30   | 14:4:30     | 14:4:30         | 14:4:30         | 23.391 ns (0 allocations: 0 bytes)      | 21.577 ns (0 allocations: 0 bytes)     |
| 10:4:30   | [2,3,4,5,6] | Int64[]         | Int64[]         | 402.547 ns (14 allocations: 1.03 KiB)   | 88.290 ns (5 allocations: 464 bytes)   |
| [1,2,3,4] | 2:4         | [2, 3, 4]       | [2, 3, 4]       | 335.186 ns (13 allocations: 976 bytes)  | 157.491 ns (7 allocations: 608 bytes)  |
| [1,2,3,4] | 2:1.0:4     | [2, 3, 4]       | [2, 3, 4]       | 345.344 ns (13 allocations: 1008 bytes) | 228.194 ns (7 allocations: 608 bytes)  |
| [1,2,3,4] | 14:4:30     | Int64[]         | Int64[]         | 332.333 ns (13 allocations: 976 bytes)  | 84.778 ns (5 allocations: 464 bytes)   |
| [1,2,3,4] | [2,3,4,5,6] | [2, 3, 4]       | [2, 3, 4]       | 333.701 ns (12 allocations: 944 bytes)  | 341.855 ns (12 allocations: 944 bytes) |

I didn't expect this PR to speed this up so much!

```
# master
julia> @btime intersect(1:1.0:4, 2:4)
  407.815 ns (16 allocations: 1.16 KiB)
3-element Vector{Float64}:
 2.0
 3.0
 4.0

# PR
julia> @btime intersect(1:1.0:4, 2:4)
  67.816 ns (2 allocations: 144 bytes)
3-element Vector{Float64}:
 2.0
 3.0
 4.0
```