@IanButterworth 2021-08-10
Is it acceptable that the return type has changed in some of the examples above?