@rfourquet 2021-08-11
> Is it acceptable that the return type has changed in some of the examples above?

Yes would be worth triaging, currently returning something of the same `eltype` as the first argument doesn't seem documented but might be a useful property? But if needed, it seems simple enough to update this PR to preserve this.