@barucden 2021-08-12
Just a note: the return type has changed for the `AbstractRange`-`AbstractVector` version too, and there's no promotion -- the vector's eltype determines the result eltype. From Ian's table:

a | b | Master return | PR Return
-- | -- | -- | --
1:1.0:4 | [2,3,4,5,6] | [2.0, 3.0, 4.0] | [2, 3, 4]

Someone might dislike this inconsistency (promotion & no promotion) between the `AbstractRange`-`AbstractRange` and `AbstractRange`-`AbstractVector` version.